# apps/gateway/docs/openapi.json
{
  "components": {
...
    "schemas": {
...
      "ErrorResponse": {
        "additionalProperties": true,
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "error",
          "message",
          "timestamp"
        ],
        "type": "object"
      },
...
      "RealtimeEventEnvelope": {
        "additionalProperties": true,
        "properties": {
          "event_id": {
            "type": "string"
          },
          "event_type": {
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object"
          },
          "route_id": {
            "type": "string"
          },
          "tenant_id": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          },
          "upstream_id": {
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "event_type",
          "occurred_at"
        ],
        "type": "object"
      },
      "RealtimeEventReplayResponse": {
        "additionalProperties": true,
        "properties": {
          "count": {
            "type": "integer"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/RealtimeEventEnvelope"
            },
            "type": "array"
          },
          "since": {
            "type": "string"
          }
        },
        "required": [
          "count",
          "events"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "apiKeyAuth": {
        "description": "API key for machine clients.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "JWT access token in the Authorization header.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Middleware-first control-plane API for RateGuard OSS.",
    "title": "RateGuard API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/events/replay": {
      "get": {
        "description": "Return a replay window of normalized realtime events for catch-up after a disconnect.",
        "operationId": "replayEvents",
        "parameters": [
          {
            "description": "Replay start timestamp in RFC3339 or Unix seconds/milliseconds",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Maximum number of events to return",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RealtimeEventReplayResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "summary": "Replay realtime events",
        "tags": [
          "system"
        ]
      }
    },
    "/api/v1/events/stream": {
      "get": {
        "description": "Open a Server-Sent Events stream for live realtime updates. Use EventSource or another SSE client to consume the response body.",
        "operationId": "streamEvents",
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Server-Sent Event stream"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "summary": "Stream realtime events",
        "tags": [
          "system"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local backend",
      "url": "http://localhost:8008"
    }
  ],
...
}